update README.md; fix small ui issues

+++ ui/ui_color.py
@@ -16,8 +16,8 @@ class UIColor:
         w = int(max(1, self.width / self.scale))
         c = (color.red(), color.green(), color.blue())
         for i in range(0, n_pnts - 1):
-            pnt1 = (points[i].x() / self.scale, points[i].y() / self.scale)
-            pnt2 = (points[i + 1].x() / self.scale, points[i + 1].y() / self.scale)
+            pnt1 = (int(points[i].x() / self.scale), int(points[i].y() / self.scale))
+            pnt2 = (int(points[i + 1].x() / self.scale), int(points[i + 1].y() / self.scale))
             cv2.line(self.img, pnt1, pnt2, c, w)
             cv2.line(self.mask, pnt1, pnt2, 255, w)
 
@@ -26,11 +26,6 @@ class UIColor:
 
     def get_constraints(self):
         return self.img, self.mask
-        # img = self.img[np.newaxis, :]
-        # mask = self.mask[np.newaxis, :]
-        # img_t = modeldef.transform(img)
-        # mask_t = modeldef.transform_mask(mask)
-        # return img_t, mask_t
 
     def update_width(self, d):
         self.width = min(20, max(1, self.width+ d))
